Backend/requirements.txt
fastapi==0.115.0
pydantic-settings==2.0.3
uvicorn[standard]==0.23.2
python-dotenv==1.0.1
SQLAlchemy==2.0.21
alembic==1.13.2
httpx==0.25.0
psycopg2-binary==2.9.8
email-validator==2.2.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
python-multipart==0.0.6
Authlib==1.3.2
fastapi-mail==1.4.0
Jinja2==3.1.2
boto3==1.35.24
requests==2.32.3
redis==4.6.0